PERA-1812 :: Moving call to MigrationManager
